test: move security warnings spec to main runner
