+++ spec/fixtures/pages/insecure-resources.html
@@ -1,6 +1,6 @@
 <html>
   <head>
-    <link rel="stylesheet" href="http://127.0.0.1:8881/save_page/test.css">
+    <link rel="stylesheet" href="/save_page/test.css">
     <script type="text/javascript">
       window.ELECTRON_ENABLE_SECURITY_WARNINGS = true
     </script>
